[ACA-4618] Move canOpenWithOffice evaluator to ACA shared
