+++ projects/adf-office-services-ext/src/public-api.ts
@@ -26,6 +26,6 @@
 export * from './lib/aos-extension.service';
 export * from './lib/actions/aos.actions';
 export * from './lib/effects/aos.effects';
-export * from './lib/evaluators';
+export * from './lib/utils';
 
 export * from './lib/aos-extension.module';
